requirements.txt
asgiref==3.5.2
Django==3.0
django-smart-selects==1.5.9
pkg_resources==0.0.0
Pillow==9.1.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2



